+++ tests/testthat/helper-initialize.R
@@ -1,3 +1,6 @@
+board_register("local", cache = tempfile(), connect = FALSE)
+board_register("packages", cache = tempfile(), connect = FALSE)
+
 test_board_is_registered <- function(board) {
   tryCatch({
     if (!board %in% board_list()) {
@@ -10,4 +13,13 @@ test_board_is_registered <- function(board) {
   }, error = function(e) {
     FALSE
   })
+}
+
+test_local_files <- function() {
+  pin_folders <- dir(board_cache_path(), recursive = TRUE)
+
+  if (!identical(pin_folders, character(0)))
+    stop("Found local files: ", paste(pin_folders, collapse = ", "))
+
+  succeed()
 }