+++ tests/testthat/test-aaa-check.R
@@ -1,0 +1,6 @@
+context("pin aaa check")
+
+test_that("tests start without local dirs", {
+  skip_on_cran()
+  test_local_files()
+})